refactor to use only one user-based helper
